+++ app/Models/Event.php
@@ -2,6 +2,7 @@
 
 namespace App\Models;
 
+use Illuminate\Database\Eloquent\Casts\Attribute;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 
@@ -16,7 +17,14 @@ class Event extends Model
         'date',
         'category_id',
         'address_id',
+        'active',
     ];
+
+    protected $casts = [
+        'active' => 'boolean'
+    ];
+
+    protected $appends = ['links'];
 
     /**
      * Obtem o endereço
@@ -30,5 +38,23 @@ class Event extends Model
     public function category()
     {
         return $this->belongsTo(\App\Models\EventCategory::class);
+    }
+
+    // accessor implementation
+    protected function links() : Attribute
+    {
+        return new Attribute(
+            get: fn () => [
+                [
+                    'rel' => 'self',
+                    'url' => "/api/events/{$this->id}",
+                ],
+                [
+                    'rel' => 'guests',
+                    'url' => "/api/events/{$this->id}/guests",
+                ],
+            ],
+            set: fn ($active) => (bool) $active,
+        );
     }
 }